requirements.txt
numpy
pandas
pyyaml
matplotlib
scikit-learn
torch
tqdm
pytest
flake8
mypy
